+++ adserver/analyzer/tasks.py
@@ -66,7 +66,7 @@ def analyze_url(url, publisher_slug, force=False):
         log.debug("Keywords from '%s': %s", backend.__name__, analyzed_keywords)
 
         analyzed_embedding = backend_instance.embedding(response)  # Can be None
-        log.debug("Embedding from '%s': %s", backend.__name__, analyzed_embedding)
+        log.debug("Embedding from '%s': %s", backend.__name__, len(analyzed_embedding))
 
         if analyzed_keywords:
             for kw in analyzed_keywords:
@@ -76,7 +76,6 @@ def analyze_url(url, publisher_slug, force=False):
             embeddings.append(analyzed_embedding)
 
     log.debug("Keywords found : %s", keywords)
-    log.debug("Embeddings found : %s", embeddings)
 
     if len(embeddings) > 1:
         log.warning("Multiple embeddings found for URL: %s", normalized_url)
